refactor: store folders
