+++ src/store/actions/actions.ts
@@ -1,3 +1,3 @@
-import { formSlice } from "./slice";
+import { formSlice } from "../slice/slice";
 
 export const { addItemToForm, removeItemFromForm, sortByProp, filterByGender, filterByInterest, returnDataAfterFiltering } = formSlice.actions;
