+++ src/types/types.ts
@@ -1,3 +1,3 @@
-import { store } from './store';
+import { store } from '../store/store';
 
 export type RootState = ReturnType<typeof store.getState>;
